requirements.txt
aiogram==3.7.0
requests==2.31.0
python-dotenv==1.0.1
Pillow>=10.0.0
reportlab>=4.0.0
